Move some things around for better abstraction
