+++ zical/Cargo.toml
@@ -20,6 +20,7 @@ path = "src/main.rs"
 [dependencies]
 diesel_migrations = "^1.4"
 rocket = "^0.4"
+tracing = "^0.1.13"
 
 [dependencies.diesel]
 version = "^1.4"
