+++ zical/src/lib.rs
@@ -11,5 +11,7 @@
 
 #[macro_use]
 extern crate diesel;
+#[macro_use]
+extern crate tracing;
 
 pub mod db;
